Fuse test_atomic_wait.hpp into P1135R6_atomic_wait_win7.
No changes other than dropping the duplicate banner and `#pragma once`.

This obviously duplicates test code, but will make removal simpler.

Note divergence: GH 4221 (merged 2024-01-17, before GH 4742 removed Win7 on 2024-06-24) added compile-only test coverage for incomplete types to `P1135R6_atomic_wait` only. There's no need to replicate that coverage.
